Make SamplesList use the new IExportDropdownItem type
